+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -6,8 +6,6 @@
 #if SENTRY_HAS_UIKIT
 #    import "MockUIScene.h"
 #    import "SentryDefaultUIViewControllerPerformanceTracker.h"
-#    import "SentryUIEventTracker.h"
-#    import "SentryUIEventTrackerTransactionMode.h"
 #    import "UIViewController+Sentry.h"
 #endif // SENTRY_HAS_UIKIT
 
@@ -108,7 +106,6 @@
 #import "SentryNSURLSessionTaskSearch.h"
 #import "SentryNetworkTracker.h"
 #import "SentryNoOpSpan.h"
-#import "SentryOptionsInternal.h"
 #import "SentryPerformanceTracker+Testing.h"
 #import "SentryPerformanceTracker.h"
 #import "SentryProfilingConditionals.h"